Fixed class structure and refactored Init class
