+++ Black Friday.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29001.49
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Black Friday", "Black Friday.csproj", "{65172524-F261-4052-BA6E-A66F84DE9FFE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Black Friday", "Black Friday.csproj", "{A33AD312-6343-443B-890D-AB71E3A25A18}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{65172524-F261-4052-BA6E-A66F84DE9FFE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{65172524-F261-4052-BA6E-A66F84DE9FFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{65172524-F261-4052-BA6E-A66F84DE9FFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{65172524-F261-4052-BA6E-A66F84DE9FFE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A33AD312-6343-443B-890D-AB71E3A25A18}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A33AD312-6343-443B-890D-AB71E3A25A18}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A33AD312-6343-443B-890D-AB71E3A25A18}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A33AD312-6343-443B-890D-AB71E3A25A18}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E5A9140C-D589-4620-9B85-7E0C825EA3FB}
+		SolutionGuid = {0D73D8D5-D161-4988-A65C-879893A0376F}
 	EndGlobalSection
 EndGlobal
